% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lysing.R
\docType{data}
\name{tengispurningar_2008_2019}
\alias{tengispurningar_2008_2019}
\title{Summuskor fyrir tengispurningar notadar 2008 og 2019}
\format{
Data frame med 6419 svorum og thremur breytum
}
\usage{
data(tengispurningar_2008_2019)
}
\description{
Gagnasettid inniheldur skor fra 6419 memendum a tengiatridum.
Nemendur eru i 10. bekk arid 2008 en 9. bekk arid 2019.
}
\details{
\itemize{
\item id er audkenni nemanda
\item prof er nafn profs (her visar thad i prof ar)
\item sig a kvarda (tekur gildi fra 0-8)
}
}
\keyword{datasets}
